backend/requirements.txt
# Backend Requirements for Countdown Timer API

flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
werkzeug==3.0.1
pytest==7.4.3
pytest-flask==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
stripe==8.0.0
supabase==2.3.0
google-generativeai==0.8.3
pyjwt==2.8.0
